+++ elide-core/src/main/java/com/yahoo/elide/core/exceptions/JsonPatchExtensionException.java
@@ -20,7 +20,21 @@ public class JsonPatchExtensionException extends HttpStatusException {
         response = Pair.of(status, errorNode);
     }
 
+    /**
+     * @deprecated use {@link #getErrorResponse()}
+     */
+    @Deprecated
     public Pair<Integer, JsonNode> getResponse() {
+        return response;
+    }
+
+    @Override
+    public Pair<Integer, JsonNode> getErrorResponse() {
+        return response;
+    }
+
+    @Override
+    public Pair<Integer, JsonNode> getVerboseErrorResponse() {
         return response;
     }
 }