+++ elide-datastore/elide-datastore-hibernate5/src/test/java/com/yahoo/elide/datastores/hibernate5/HibernateDataStoreSupplier.java
@@ -7,6 +7,7 @@ package com.yahoo.elide.datastores.hibernate5;
 
 import com.yahoo.elide.core.DataStore;
 import com.yahoo.elide.models.generics.Manager;
+import com.yahoo.elide.models.triggers.Invoice;
 import com.yahoo.elide.utils.ClassScanner;
 
 import example.Filtered;
@@ -53,6 +54,8 @@ public class HibernateDataStoreSupplier implements Supplier<DataStore> {
             ClassScanner.getAnnotatedClasses(Parent.class.getPackage(), Entity.class)
                     .forEach(metadataSources::addAnnotatedClass);
             ClassScanner.getAnnotatedClasses(Manager.class.getPackage(), Entity.class)
+                    .forEach(metadataSources::addAnnotatedClass);
+            ClassScanner.getAnnotatedClasses(Invoice.class.getPackage(), Entity.class)
                     .forEach(metadataSources::addAnnotatedClass);
         } catch (MappingException e) {
             throw new RuntimeException(e);