+++ changelog.md
@@ -3,6 +3,9 @@
 **Fixes**
  * Issue#744 Better error handling for mismatched method in Lifecycle and additional test
 
+**Features**
+ * Let custom Elide wrapper process JsonPatchExtensionException
+
 ## 4.3.2
 **Fixes**
  * Issue#754